+++ Src/FluentAssertions/Equivalency/Tracing/InternalTraceWriter.cs
@@ -1,0 +1,5 @@
+namespace FluentAssertions.Equivalency.Tracing;
+
+internal sealed class InternalTraceWriter : StringBuilderTraceWriter
+{
+}